+++ packages/Webkul/Core/src/Http/Middleware/SecureHeaders.php
@@ -43,6 +43,7 @@ class SecureHeaders
         $response->headers->set('X-XSS-Protection', '1; mode=block');
         $response->headers->set('X-Frame-Options', 'DENY');
         $response->headers->set('Strict-Transport-Security', 'max-age=31536000; includeSubDomains');
+        $response->headers->set('X-Built-With', 'Bagisto');
     }
 
     /**