+++ packages/Webkul/Razorpay/src/Contracts/RazorpayTransaction.php
@@ -1,0 +1,5 @@
+<?php
+
+namespace Webkul\Razorpay\Contracts;
+
+interface RazorpayTransaction {}